refactor(wsdl): move diagnostics ownership out of cli helpers
Extract generic diagnostics lifecycle helpers into cli/helpers/diagnostics,
move WSDL-specific diagnostics key and wrapper under manifests/wsdl,
and reuse the WSDL-owned key in schema diagnostics storage.
