+++ spinta/manifests/wsdl/__init__.py
@@ -1,3 +1,9 @@
+from .diagnostics import WSDL_SCHEMA_DIAGNOSTICS_KEY, wsdl_schema_diagnostics
 from .helpers import is_wsdl_path, normalize_wsdl_path
 
-__all__ = ["is_wsdl_path", "normalize_wsdl_path"]
+__all__ = [
+	"is_wsdl_path",
+	"normalize_wsdl_path",
+	"WSDL_SCHEMA_DIAGNOSTICS_KEY",
+	"wsdl_schema_diagnostics",
+]
