+++ spinta/manifests/wsdl/commands/diagnostics.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 
 from spinta import commands
+from spinta.cli.helpers.diagnostics import get_diagnostics_storage
 from spinta.components import Context
 from spinta.manifests.wsdl.components import WsdlManifest
+from spinta.manifests.wsdl.diagnostics import WSDL_SCHEMA_DIAGNOSTICS_KEY
 
 
 @commands.store_schema_diagnostics.register(Context, WsdlManifest, str, list)
@@ -22,9 +24,6 @@ def store_schema_diagnostics(
     if not diagnostics:
         return
 
-    if not context.has("wsdl.schema_diagnostics", value=True):
-        context.set("wsdl.schema_diagnostics", [])
-
-    stored = context.get("wsdl.schema_diagnostics")
+    stored = get_diagnostics_storage(context, WSDL_SCHEMA_DIAGNOSTICS_KEY)
     for diagnostic in diagnostics:
         stored.append((normalized_path, str(diagnostic)))
